Move files and add files field
