+++ packages/eth-rpc-cache/package.json
@@ -7,6 +7,11 @@
     "eth",
     "rpc"
   ],
+  "files": [
+    "dist",
+    "src"
+  ],
+  "main": "dist/index.js",
   "bugs": "https://github.com/bloq/eth-packages/issues",
   "license": "MIT",
   "author": {
