+++ packages/eth-rpc-cache/src/error.ts
