+++ packages/eth-rpc-cache/src/strategies/permanent.ts
