+++ packages/eth-rpc-cache/src/utils/clone.ts
